trwe

+++ app/src/main/java/com/softtek/lai/module/laicheng/model/BleMainData.java
@@ -161,7 +161,7 @@ public class BleMainData implements Serializable{
         private String BirthDate;
         private int Age;
         private int Gender;
-        private int Height;
+        private float Height;
         private String PhoneNo;
 
         public String getName() {
@@ -196,11 +196,11 @@ public class BleMainData implements Serializable{
             this.Gender = Gender;
         }
 
-        public int getHeight() {
+        public float getHeight() {
             return Height;
         }
 
-        public void setHeight(int Height) {
+        public void setHeight(float Height) {
             this.Height = Height;
         }
